Full project refactor
* Completely restructured the project, shifting focus away from cats and removing all unnecessary features.

Added a new `BoardController`, which will be used for the News page. Implemented functionality for users to leave comments and react to both site updates and other comments.

Updated avatar storage on the backend: avatars are now saved as an FontAwesome icon name plus a color code instead of base64 images.